Remove home direction to share page
Switch share page as index
Disable redirection through cookie (much accessibility)
