+++ pages/index.php
@@ -16,75 +16,34 @@ limitations under the License.
 Code source hosted on https://github.com/nicolabricot/MoodPicker
 */
 
+use Utils\Cookie;
+use Utils\Session;
 use Picker\Mood;
 use Picker\MoodLevel;
-use Utils\Session;
-use Utils\Cookie;
-use Utils\TextHelper;
 
-// if first visit and not voted, ask mood
-if (! Session::Exists('welcome') && ! Cookie::Exists('voted')) {
-    Session::Add('welcome');
-    header('Location: ./share');
-    exit();
-}
-
-$moods = Mood::DayMoods();
-$bads = array();
-$goods = array();
-foreach ($moods as $m) {
-    if ($m->getMood() == MoodLevel::GOOD)
-        $goods[] = $m;
-    else
-        $bads[] = $m;
-}
-
-// results of the day
-$this->assign('moods', $moods);
-if (count($moods) > 0) {
-    $this->assign('goods_percentage', count($goods) * 100 / count($moods));
-    $this->assign('bads_percentage', count($bads) * 100 / count($moods));
-    $s = "
-    $(function(){
-        var data = [
-            {
-                value: ".count($bads).",
-                color: $('#color_picker .progress-bar-danger').css('background-color'),
-                label: 'Bad Mood'
-            },
-            {
-                value: ".count($goods).",
-                color: $('#color_picker .progress-bar-success').css('background-color'),
-                label: 'Good Mood'
-            }
-        ]; 
-    ";
-    if (count($moods) > 0) {
-        $s .= "
-            new Chart(document.getElementById('dayChart').getContext('2d')).Doughnut(data, { 
-                animation: true, animationEasing: 'linear', animationSteps: 25,
-                tooltipFontFamily: $('body').css('font-family'),
-                tooltipFontSize: 12,
-                segmentShowStroke: false,
-                percentageInnerCutout : 60
-            });
-        ";
+if (isset($_POST['mood']) && $this->acceptToken()) {
+    $_POST['mood'] = $_POST['mood']+0;
+    
+    if (! MoodLevel::isValidValue($_POST['mood']))
+        $this->errorPage('Invalid value', 'The given value for your current mood is unknow.');
+    
+    if (!Cookie::Exists('voted') && !Session::Exists('voted')) {
+        $m = new Mood($_POST['mood'] , time(), $_SERVER['REMOTE_ADDR']);
+        $m->save();
+        Cookie::add('voted', true, Cookie::HOUR*2);
+        Session::add('voted', true);
+        
+        header('Location: ./review');
+        exit();
     }
-    $s .= "});";
-    $this->register('script', TextHelper::removeLineBreak($s));
+    
+    $this->errorPage('Already voted', 'An entry has already been enregistred from your computer. <br />You have to wait some times before submitting an other mood.');
+    
 }
-
-// stats of the month
-$month = Mood::CountMonthMoods();
-$month_goods = Mood::CountMonthMoods(NULL, NULL, MoodLevel::GOOD);
-$month_bads = $month - $month_goods;
-if (empty($month)) { $month = 1; }
-$this->assign('month_goods_percentage', $month_goods * 100 / $month);
-$this->assign('month_bads_percentage', $month_bads * 100 / $month);
-$this->assign('month_goods', $month_goods);
-$this->assign('month_bads', $month_bads);
-
-// javascript dependancy
-$this->register('script_file', 'chart.min.js');
+else {
+    $this->getToken();
+    $this->assign('good', MoodLevel::GOOD);
+    $this->assign('bad', MoodLevel::BAD);
+}
 
 ?>
