+++ templates/review.html
@@ -15,7 +15,7 @@
 	
         <h2>Today’s review <small>{function="date('j F Y')"}</small></h2>
         {if="empty($moods)"}
-        <div class="row"><div class="col-md-6 col-md-offset-3"><div class="alert alert-warning text-center" role="alert">No one has submitted his mood today :-(<br /><a href="share">Submit yours!</a></div></div></div>
+        <div class="row"><div class="col-md-6 col-md-offset-3"><div class="alert alert-warning text-center" role="alert">No one has submitted his mood today :-(<br /><a href="{$URL}">Submit yours!</a></div></div></div>
         {else}
     	<div class="media espace-top">
     	    <div class="pull-left">
